fduy

+++ src/ejercicio1/Complejo.java
@@ -1,38 +1,38 @@
 package ejercicio1;
 
 public class Complejo {
-	private float real;
-	private float imaginario;
+	private double real;
+	private double imaginario;
 	
 	public Complejo(){
 		real=0;
 		imaginario=0;
 	}
 	
-	public Complejo(float real,float imaginario){
+	public Complejo(double real,double imaginario){
 		this.real=real;
 		this.imaginario=imaginario;
 	}
 	
-	public void setReal(float real){
+	public void setReal(double real){
 		this.real=real;
 	}
 	
-	public void setImaginario(float imaginario){
+	public void setImaginario(double imaginario){
 		this.imaginario=imaginario;
 	}
 	
-	public float getReal(){
+	public double getReal(){
 		return real;
 	}
 	
-	public float getImaginario(){
+	public double getImaginario(){
 		return imaginario;
 	}
 	
 	
 	public Complejo suma(Complejo numero){
-		float sumareal,sumaimaginario;
+		double sumareal,sumaimaginario;
 		sumareal=this.real+numero.getReal();
 		sumaimaginario=this.imaginario+numero.getImaginario();
 		
@@ -40,7 +40,7 @@ public class Complejo {
 	}
 	
 	public Complejo resta(Complejo numero){
-		float sumareal,sumaimaginario;
+		double sumareal,sumaimaginario;
 		sumareal=this.real-numero.getReal();
 		sumaimaginario=this.imaginario-numero.getImaginario();
 		
@@ -48,7 +48,7 @@ public class Complejo {
 	}
 	
 	public Complejo multiplicacion(Complejo numero){
-		float sumareal,sumaimaginario;
+		double sumareal,sumaimaginario;
 		sumareal=this.real*numero.getReal();
 		sumaimaginario=this.imaginario*numero.getImaginario();
 		
@@ -56,7 +56,7 @@ public class Complejo {
 	}
 	
 	public Complejo division(Complejo numero){
-		float sumareal,sumaimaginario;
+		double sumareal,sumaimaginario;
 		sumareal=this.real/numero.getReal();
 		sumaimaginario=this.imaginario/numero.getImaginario();
